requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
cryptography==42.0.8
Django==5.0.6
django-allauth==0.63.3
django-crispy-forms==2.2
idna==3.7
jwt==1.3.1
pycparser==2.22
PyJWT==2.8.0
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
